FileDistributor/requirements.txt
kafka-python==2.0.5
peewee==3.17.9
psycopg2==2.9.10